restructure wikipedia references
I'm not sure about the protocol around referencing WP in
internet-drafts.

But if we're going to reference WP, we should at least do it with
formatting that kramdown-rfc2629 will understand. :)

WP pages seem much more ephemeral than RFCs.  Maybe we also need to
include a date ("date retrieved"?) in the reference section?