Changed wording
This sentence "that the compromise made in design of the protocol has resulted in a technology which failed to technical encode the freedom of expression goals it was designed to promote." was creating confusion as it seemed to suggest that a new technology has not been adopted because of an IP design [mis]feature. It has been rewritten to prevent this misreading.