added language on human rights and new gtlds
This development is seen as a double-edged sword in terms of human rights. On the one hand individuals have argued that it limits freedom of expression when a gTLd has a restricted policy on registration and rules on hate speech content. On the other hand others have argued that it enables particular communities to build safer places of association, and as such improves the right of freedom of association. These new gTLds can function as private spaces for certain communities, like for instance those coalescing around applications like .gay.