Updated language
on torture and extrajudicial killings, to focus on the fact that unfortunately it is not only totalitarian regimes that practice torture:
regimes that consider themselves to be democratic do so also.  And regimes
that consider themselves to be democratic engage in extrajudicial killing or
detention, see:

 https://en.wikipedia.org/wiki/Extrajudicial_killing

 https://en.wikipedia.org/wiki/Administrative_detention

As per the suggestion of R. Hill.