Largely semantics
Changing citizens to individuals: an individual belongs to him or herself, while citizenship is something granted or appended to the individual by a government or a state. Individual is just more inclusive and better reflective of the message of the message conveyed, in my view.

Removing references to "protocol police": As a political catch-phrase, it's unfortunate because it supports the notion that such a thing as a "protocol police" could or would exist. It's a defensive statement that doesn't serve help in clarifying the research topic: on the contrary, it would be an interesting research topic in itself to investigate the emergence and propagation of this meme. 

Removing "should not": It doesn't need to be clarified what the IETF should not do, since its inability to do that thing is already inherent to what the IETF is ("voluntary" standards would not in fact be voluntary if they could be involuntary).